pyproject.toml
[tool.poetry]
name = "kover-bot"
version = "0.1.0"
description = ""
authors = ["Marat <esquonk@gmail.com>"]
readme = "README.md"
packages = [{ include = "kover_bot" }]

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "^4.10.0"
python-telegram-bot = "^20.0b0"
pytz = "*"
aiohttp = "^3"
reactivex = "^4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
